requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
pymongo
certifi
pymongo[srv] == 3.6

-e .